Add the parser and a function to execute a parsed query
The function parseQuery(query, params) parse a query whith his parameters. The querry is to be used in an aggregate function.
It takes two parameters:
- A string which is a mongo query
Exemple : [{\"$match\": {\"founded_year\": {\"$lt\": ???}}}]
Please note that if you write it manually you've got to escape the ", otherwise it will be automatic (from a html form for exemple).

- An array of parameters for exemple [2000] for the previous query. Note that ["2000"] would yield the same result.

If you have to add a string as a parameter (note a number) you have to write it this way : \"...\"
So when the query will be parse the content of the string will be placed in between the two apostrophes.